Air Hockey Game
Demo for Win32 systems, based on the "Win32 Game Engine". Supports
Kinect sensor, both through OpenNI and Microsoft's Kinect SDK. The
executable files are included.

You will need to have installed and updated DirectX to run the demo. In
order to use the Kinect sensor to play, the appropriate runtime package
(either OpenNI or Kinect SDK) is required and must be installed.